Cleaned and updated utm courses scraper
