+++ .gitignore
@@ -2,3 +2,8 @@ __pycache__
 utils/__pycache__
 .DS_Store
 **/.DS_Store
+.idea/
+.vscode/
+.mypy_cache/3.6
+
+
