--- utils/__init__.py
@@ -1,2 +1,0 @@
-from .scraper import Scraper
-from .layers import LayersScraper
